requirements.txt
rich>=14.1.0
requests>=2.32.4
python-dotenv>=1.1.1
coverage>=7.10.2